train.py
import os
import yaml
import torch
from ultralytics import YOLO
import argparse
from pathlib import Path

def download_coco_dataset(data_dir="datasets"):
    """
    Download and prepare the COCO dataset for training.
    In a real implementation, you would download the actual COCO dataset.
    For this demo, we'll create a simplified version.
    """
    os.makedirs(data_dir, exist_ok=True)
    coco_dir = os.path.join(data_dir, "coco")
    os.makedirs(coco_dir, exist_ok=True)
    
    # Create dataset directories
    for split in ["train", "val"]:
        os.makedirs(os.path.join(coco_dir, split, "images"), exist_ok=True)
        os.makedirs(os.path.join(coco_dir, split, "labels"), exist_ok=True)
    
    # Create YAML file for COCO dataset
    coco_yaml = {
        "path": coco_dir,
        "train": "train/images",
        "val": "val/images",
        "names": {
            0: "person", 1: "bicycle", 2: "car", 3: "motorcycle", 4: "airplane",
            5: "bus", 6: "train", 7: "truck", 8: "boat", 9: "traffic light",
            # ... more classes
        }
    }
    
    with open(os.path.join(data_dir, "coco.yaml"), "w") as f:
        yaml.dump(coco_yaml, f)
    
    print(f"COCO dataset structure created at {coco_dir}")
    print("In a real implementation, you would download the actual COCO dataset.")
    
    return os.path.join(data_dir, "coco.yaml")

def train_yolo_model(
    data_yaml,
    model_size="n",  # n, s, m, l, x
    epochs=10,
    batch_size=16,
    img_size=640,
    project="models",
    name="train"
):
    """
    Train a YOLOv8 model on the specified dataset.
    
    Args:
        data_yaml: Path to the dataset YAML file
        model_size: YOLOv8 model size (n, s, m, l, x)
        epochs: Number of training epochs
        batch_size: Batch size
        img_size: Image size
        project: Directory to save results
        name: Name of the experiment
    
    Returns:
        Path to the best trained model
    """
    # Load a pre-trained YOLOv8 model
    model = YOLO(f"yolov8{model_size}.pt")
    
    # Train the model
    results = model.train(
        data=data_yaml,
        epochs=epochs,
        batch=batch_size,
        imgsz=img_size,
        project=project,
        name=name
    )
    
    # Get the path to the best model
    best_model_path = os.path.join(project, name, "weights", "best.pt")
    
    return best_model_path

def main():
    parser = argparse.ArgumentParser(description="Train YOLOv8 model for object detection")
    parser.add_argument("--data", type=str, default="", help="Path to data YAML file")
    parser.add_argument("--model-size", type=str, default="n", choices=["n", "s", "m", "l", "x"], help="YOLOv8 model size")
    parser.add_argument("--epochs", type=int, default=10, help="Number of training epochs")
    parser.add_argument("--batch-size", type=int, default=16, help="Batch size")
    parser.add_argument("--img-size", type=int, default=640, help="Image size")
    parser.add_argument("--project", type=str, default="models", help="Directory to save results")
    parser.add_argument("--name", type=str, default="train", help="Name of the experiment")
    parser.add_argument("--download-coco", action="store_true", help="Download COCO dataset")
    
    args = parser.parse_args()
    
    # Download COCO dataset if requested
    if args.download_coco:
        data_yaml = download_coco_dataset()
    else:
        data_yaml = args.data
    
    # Check if data YAML file exists
    if not data_yaml:
        print("Error: No data YAML file specified. Use --data or --download-coco")
        return
    
    # Train the model
    print(f"Training YOLOv8{args.model_size} model on {data_yaml} for {args.epochs} epochs")
    best_model_path = train_yolo_model(
        data_yaml=data_yaml,
        model_size=args.model_size,
        epochs=args.epochs,
        batch_size=args.batch_size,
        img_size=args.img_size,
        project=args.project,
        name=args.name
    )
    
    print(f"Training completed! Best model saved to {best_model_path}")

if __name__ == "__main__":
    main()